Move history file to bash directory
When importing history gist as a subtree, I had to put the file in
separate directory (that’s how subtrees work), but I want to have it
next to the other files. Fortunately git is smart enough that any
subsequent merges should apply changes to the file in the new location.