History: keep default location & don't set HISTTIIMEFORMAT
It seems that some combination of multiple history files with mixed
format (some using command timestamps and some not) causes bash to exit
on loading history. The simplest solution is not to mess with timestamp
settings - let the user set them herself.

Also, the principle of least surprise suggests that HISTFILE shouldn't
be changed if not necessary.